+++ lib/common/htmltable.c
@@ -1094,7 +1094,7 @@ static int size_html_img(htmlimg_t * img, htmlenv_t * env)
 	agerr(AGERR, "No or improper image file=\"%s\"\n", img->src);
     } else {
 	rv = 0;
-	GD_has_images(env->g) = TRUE;
+	GD_has_images(env->g) = true;
     }
 
     B2BF(b, img->box);
@@ -2025,7 +2025,7 @@ int make_html_label(void *obj, textlabel_t * lp)
 	agxbuf xb;
 	unsigned char buf[SMALLBUF];
 	agxbinit(&xb, SMALLBUF, buf);
-	lp->html = FALSE;
+	lp->html = false;
 	lp->text = strdup(nameOf(obj, &xb));
 	switch (lp->charset) {
 	case CHAR_LATIN1: