+++ lib/dotgen/position.c
@@ -1116,10 +1116,10 @@ int ports_eq(edge_t * e, edge_t * f)
     return ((ED_head_port(e).defined == ED_head_port(f).defined)
 	    && (((ED_head_port(e).p.x == ED_head_port(f).p.x) &&
 		 (ED_head_port(e).p.y == ED_head_port(f).p.y))
-		|| (ED_head_port(e).defined == FALSE))
+		|| !ED_head_port(e).defined)
 	    && (((ED_tail_port(e).p.x == ED_tail_port(f).p.x) &&
 		 (ED_tail_port(e).p.y == ED_tail_port(f).p.y))
-		|| (ED_tail_port(e).defined == FALSE))
+		|| !ED_tail_port(e).defined)
 	);
 }